VirtualDisplays (v2): Use a per-UID thread for VDs
Instead of per-VD threads like before, we now use a single thread per
UID. This CL does not supply the UIDs, which will be woven through SF
and DM in other CLs.

To make this possible, we add a VirtualDisplayThreadManager singleton who keeps
track of each VirtualDisplayThread and makes it available or constructs
it. We keep a reference count of VirtualDisplays that are using each
thread and will destroy it when there are no more displays.

Bug: 340933138
Flag: com.android.graphics.surfaceflinger.flags.wb_virtualdisplay2
Test: atest, presubmit, new tests

Change-Id: Ie6aa7338f51ba43dd303d10af270925422f3fc8d