SF: Use specified pacesetterId in designatePacesetterDisplay
Currently, if the pacesetter_selection flag is enabled, Scheduler
ignores any pacesetterId argument passed into
Scheduler::designatePacesetterDisplay() and invokes
selectPacesetterDisplay() to select a new pacesetter. Same behavior
is present in Scheduler::promotePacesetterDisplayLocked().

This CL changes the condition for invoking selectPacesetterDisplay() in
both methods to only when a pacesetterId is not specified. This change
allows the caller to specify the pacesetter display.

Flag: com.android.graphics.surfaceflinger.flags.pacesetter_selection
Bug: 389983418
Test: atest libsurfaceflinger_unittests with and without the flag
Change-Id: If37b633d05f4be69615a44d2c3fb01502cf23d67