IPCThreadState: remove spurious call to readInt32 in BC_FROZEN_BINDER
The actual value of isFrozen is provided as part of the
binder_frozen_state_info that we already read in the readInPlace() call.
The only reason that reading a spurious i32 doesn't break anything is
that the buffer is always empty in this scenario because the kernel
driver stops after delivering BC_FROZEN_BINDER under the assumption that
they might result in transactions.

Test: binderLibTest
Flag: EXEMPT bugfix
Change-Id: Iae2add531ddfffedb91ad9dd42efd338c9b434f8