SF: Designate pacesetter display in powerOffDuringConcurrentModeSet
DisplayModeSwitchingTest#powerOffDuringConcurrentModeSet relies on the
internal display (the first front internal display) being the
pacesetter. This is because the outer display (non-pacesetter) is turned
off during the test, and the pacesetter is thus not switched and
VsyncModulator does not receive a cancel refresh rate change. If the
outer display was the pacesetter/front internal display, the
ModeSwitchingTo() matcher would fail.

With the pacesetter_selection flag enabled, the pacesetter is the
powered on display with the highest refresh rate, which is the outer
display for this test case. So if the outer display is turned off, the
pacesetter would move and vsync modulator would have its refresh rate
change cancelled. To be consistent with the original test case, this CL
manually sets the pacesetter to be the inner display.

Flag: EXEMPT - test only
Bug: 389983418
Test: atest libsurfaceflinger_unittests with and without the flag
Change-Id: Ic021695eace9cd2aa374037a2b06c8233dac5909