SF: Wire up mock enableLayerCachingTexturePool
If the pacesetter_selection flag is disabled, the layer caching texture
pool for the front internal display is enabled/disabled through
SF::onNewFrontInternalDisplay().

However, when the pacesetter_selection flag is enabled, the pool is
enabled for the pacesetter display and triggered by Scheduler's call to
ISchedulerCallback::enableLayerCachingTexturePool(), which SF
implements.

However, in TestableSurfaceFlinger, the ISchedulerCallback
implementation that is passed into the Scheduler is a mock. So if a call
to enable layer pool is invoked here, one would need to set up the mock
behavior.

This CL sets up a mock behavior in
FoldableTest#layerCachingTexturePoolOnFrontInternal to just invoke the
implementation in SurfaceFlinger.

Flag: EXEMPT test fix
Bug: 389983418
Test: atest libsurfaceflinger_unittests with and without the flag
Change-Id: Ib3acb1d1b7ab52c3e87e141be74711987b4e3a09