Remove AddAccessorEmpty test case in binderAllocationLimits
The ALOGE sometimes doesn't cause an allocation and that makes this test
case flaky.

It's not a very important test case, as it is only verifying no
allocations occur(outside of that log) in an uncomon failure case when
trying to inject Accessors with an empty list of instances.

Flag: TEST_ONLY
Bug: 427290547
Test: atest binderAllocationLimits
Change-Id: I96b06d375c733fe63216ad1b76148a2ac19cacbc