SF: use a lower time to resync based on transaction
We currenly use 750ms threshold for resync to vsync based.
This value is too high for an accurate prediction, but we use
it to prevent severe power drain on apps use Choreographer without
rendering (like the case of a leaking animators). For transactions,
since we anyway wakeup SF main thread, there is less changes for a
misbehaved to wakeup SF without the intention to render.

Bug: 425994753
Test: ABTD (see bug for details)
Flag: com.android.graphics.surfaceflinger.flags.resync_on_tx
Change-Id: Icdd2c6755708e5ad934b7e240ddbceb20d7d8e23