Send the ModeRejected event only if EventRegsitration is set
The modeRejected event sent by Surfaceflinger was consumed by all
the listeners (Display Event Receiver and Choreographer) causing
a crash in Choreographer for this event since dispatchModeRejected()
in Choreographer is not registered.

This patch adds a Event registration flag corresponding to the
mode rejected event and only the listeners where this flag
is set will receive this modeRejected event sent by
SF. SO in this case only DER receives this event.

Test: manual test
Bug: 409640463
Flag: EXEMPT bug fix

Change-Id: Id006a175bbf8dc9e1d19d6b42f66783aae9708d9
Signed-off-by: Manasi Navare <navaremanasi@google.com>