input: Remove edge flags from internals
For a long time, these have been unused by the input system (which
always sets them to NONE), so we can get rid of them from the input
internals. However, consumers of input events sometimes use
MotionEvents' edge flag fields to store their own metadata about the
event. (For example, Launcher sets a bit in the edge flags to denote
touches that were dispatched to the navigation bar.) This means that we
have to keep storing them in the MotionEvent, as opposed to changing
MotionEvent::getEdgeFlags to always return NONE.

Test: check that touchscreen and touchpad input still works, including
      for system navigation gestures
Test: $ m checkinput
Test: $ atest --host inputflinger_tests
Bug: 321101159
Flag: EXEMPT refactor

Change-Id: Id8eaed945875f5871cb57c1924e9642dbb8f8e59