Print AIDL 'char' values as numbers.
I found only one case where 'char' or 'char[]' store character
data. In the vast majority of cases, 'char' fields and parameters
are used as 16-bit unsigned numeric values and it's not useful to
print them as characters.

Using 'char' to represent a Unicode character in an AIDL interface
seems incorrect anyway. The 'char' type in AIDL is a 16-bit integer
that only fits a single UTF-16 code unit, so it can only represent
BMP code points, and a single code point does not match end users'
idea of a "character". The only correct way to represent text in
AIDL is to use `String`.

Bug: 244494451
Test: presubmit
Flag: EXEMPT infeasible
Change-Id: I51e3d804a5302f4ded6a5450a4d18d3f840212f0