Update MouriMap to use SkShader::makeWithWorkingColorSpace()
This removes all uses of per-pixel to/fromLinearSrgb and explicitly has
the temporary offscreen images created with a linear sRGB color space
(so that the input is mapped automatically to this space when sampling).
Then the final tonemapping step is wrapped with
makeWithWorkingColorSpace(linearSRGB) to also automatically map the
input to linear, and to have a no-op conversion for the offscreen
buffers. Its output is then automatically mapped to whatever dst color
space the returned shader is rendered into.

Bug: b/426601394
Flag: EXEMPT bug fix
Change-Id: I62810ce8f8ded9c44de5182d79c123d112d9d229