Update GainMap to assume it is rendering a linear surface
SkShaders return color values in the dst color space, so `sdr` and `hdr`
children of the gainmap return linear values if the gainmap is rendered
into a linear surface.

The gainmap shader could be made more flexible by wrapping it in a call
to SkShader::makeWithWorkingColorSpace(linear), but it is only used in
one place so I opted for documenting the requirements of the shader and
how either approach could work. Since the gainmap shader is the final
effect of the composition in the current usecase, using the dst
colorspace to handle conversion keeps things simplest.

Flag: EXEMPT bug fix
Bug: 426601394
Change-Id: I49596e6a161b509deaccbdc31dd2a342ef771862