Add API to binder to set min threads
If an RpcServer has a rootObject that requires a minimum number of
threads that is less than its mMaxThreads, it will overwrite the
mMaxThreads with this minimum thread number.

This is intended to avoid the need to expose RpcServer and its
implementation details in higher layers like ndk and rust.

This is only used for interactions with RpcServers right now.

Flag: EXEMPT adding new API, not using it
Test: atest binderRpcTest
Bug: 430314963
Change-Id: I3bb674d4c68483243c4a4b89cbb4c827b6c59e70