Add CornerRadiiCallback to BLASTBufferQueue
This adds a callback the client can register to receive
corner radii computed for the layer so that rounding can be done
in the client. The effective corner radii are computed when
the corner radius is updated and populated in SurfaceStats which are
sent with TransactionCallback. Also adds a clientCrop parameter to
setClientDrawnCornerRadius call for client to pass the crop they are
using for rounding

Bug: 372620947
Flag:com.android.graphics.surfaceflinger.flags.set_client_drawn_corner_radii
Test: presubmit; atest SurfaceFlinger_test libsurfaceflinger_unittest libgui_test
Change-Id: I10be3cfd0d55f7bfe972cbfa450da3a283ff85a4