Add System Content Priority API
The API is used to configure the importance of the layer's content from
the system's perspective. A higher priority value will give the layer
preferred access to limited resource.

The priority value is supposed to be set by system server via
SurfaceControl#Transaction. Then it propergates to LayerState and
LayerSnapshot.

Bug: 431285755
Test: presubmit
Flag: EXEMPT API not in use yet.

Change-Id: I2dfb5cc9162c1453fdbe6db002014acc63b1fee3