DisplayIdentification: fix EDID parsing
There is a bug where instead of considering the first preferred DTD
timing, DisplayIdentification goes through all the DTD timings
overwriting the preferred timing.
This was reported on Android TV where a the NTS DISPLAY-CM-ATV-TC1
started failing, so the edid for that test was also added to unit tests
for DisplayIdentification.

Flag: EXEMPT bugfix
Bug: 421350173
Test: DisplayIdentification_test
Change-Id: I295579cf0beebe7f53b01b20eeb75ca7fa9afec0