SF: Loosen Scheduler::chooseRefreshRateForContent() skip condition
Currently Scheduler::chooseRefreshRateForContent() skips if the
pacesetter display cannot switch seamlessly. With PDARR, the pacesetter
may not correspond to the internal display, so this may not be true.

Loosen the skip condition to only skip chooseRefreshRateForContent()
when none of the powered on display can switch.

Flag: com.android.graphics.surfaceflinger.flags.follower_arbitrary_refresh_rate_selection
Bug: 371584290
Test: New SchedulerTest
Change-Id: I2adfd14cffbb9ce3063c9802c0b853f77d9bc5f9