libtimeinstate: Skip over missing uids
AllUidTimeInStateMonotonic and AllUidConcurrentTimesMonotonic are test
cases that get the CPU frequency time map for all of the UIDs, sleep,
and then get the map again. Previously, we would fail the test case if
one of the UIDs from map1 was missing in map2, but this creates problems
when UIDs are explicitly removed by the KernelCpuUidTimeReader, for
example after `dumpsys batterystats --sync` is called. Since missing
a particular UID in the map does not necessarily indicate a failure,
skip over that UID in the test.

Bug: 422280686
Bug: 413162440
Test: treehugger
Flag: EXEMPT test only
Change-Id: Idcc18802bda168cb96108f2290e47acedb66eaf7