Fix incorrect assumptions about ashmem_valid()
This test assumes that ashmem_valid() tests an fd to see if it refers
to /dev/ashmem*. This is not correct; ashmem_valid() tests an fd to
ensure that it is compatible with the API defined by libcutils, so it
will return true for any fd that references /dev/ashmem* or is a memfd.

Given that some of the tests explicitly expect memfds, use a more
rigorous function to check if a given fd is a memfd.

Bug: 433411899
Flag: EXEMPT bugfix
Change-Id: I33dd4058a79f30736f8cf2e879073f8c3e94894c
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>