SF: Force GPU composition of slower follower displays
Buffers may prematurely release if they are device composited on a
display slower than the composition rate as set by the pacesetter's
refresh rate. To work around this, this CL forces client composition of
all layers presenting on a display that is configured on a lower refresh
rate than the pacesetter. Existing member of CompositionRefreshArgs used
to force client composition on *all displays*
(devOptForceClientComposition) is changed to be a list of displays
forced to client composite.

Refer to go/pdarr-buffer for more details.

Flag: com.android.graphics.surfaceflinger.flags.force_slower_follower_gpu_composition
Bug: 435235974
Test: Existing OutputTest coverage on forced client composition
Test: Manually verified on comet w/ connected display via dumpsys SF
Change-Id: Ifdfb60b00068d432afe9e35b99c5577ec91b6f95