RPC Binder: don't chunk by default
This change raises the default chunk size to the transaction limit size.
So, we will never break transaction apart unless we are hitting a low
memory change.

Originally, this chunk limit was put in-place due to some vsock kernel
issues that caused large packets to silently drop data. However, as this
has been fixed, we are restoring the limit back down.

Bug: 424526253
Bug: 416772676
Test: binderRpcTest
Flag: EXEMPT . (CL in teamfood)
Change-Id: Ia2eba7cb96c0224f4a458ee8a00ff90254ada7db