Add the concept of a MergeableHierarchy
A MergeableHierarchy is a new concept that tracks LayerHierarchies
that can be transformed into a simpler set of LayerHierarchies

This patch also introduces the concept of an Accumulator, which will
walk down the hierarchy when computing LayerSnapshots to make the
decision whether to add a LayerHierarchy into a MergeableHierarchy

Currently, all LayerHierarchies will always be added into a
MergeableHierarchy, but future patches will introduce heuristics such
as layer inactivity to selectively add layers.

We also add information into dumpsys, so that we can verify how
MergeableHierarchies are being created.

Bug: 414820185
Flag: com.android.graphics.surfaceflinger.flags.frontend_caching_v0
Test: dumpsys SurfaceFlinger
Change-Id: I8919b408016c19b9f9ab978ccd745773e0a801ab