Trace individual GrAlloc allocations in process
We have a common pain point where we can't determine the reasons for
memory growth. We can commonly guess that it might be new gralloc
allocations of e.g. surfaces, but it's hard to figure that out from
traces.

This CL adds event and buffer counter track which will make it
significantly easier to detect overallocation of memory in lab and root
cause regressions.

Bug: 435366750
Flag: EXEMPT tracing only
Test: grabbed a trace from a cheetah
Change-Id: I6972ceaef1365b9b480948144001759d63d3327d