SF: Fix port duplication check in legacy mode
In legacy mode, a port duplication check incorrectly prevents an
external display from connecting if it also lacks a valid EDID.

When a display without an EDID is hot-plugged, the system receives an
uninitialized port number that defaults to 0. The duplication check
then fails, as this conflicts with the internal display's predefined
legacy port ID, which is also 0.

This change resolves the issue by skipping the port duplication check
when the system is in legacy mode. This is safe because legacy mode is
hardcoded to support a maximum of two displays, inherently ignoring any
others and thus preventing actual port conflicts.

Bug: 431089447
Bug: 441589571
Flag: EXEMPT BUGFIX
Test: presubmit
Change-Id: Ibc4e15a04126af95478981be2b50400c20a3d863