Notify caller when motion could not be split
Currently, the splitting functionality may fail. The logic is
implemented in libinput, where limited context about the data is
available.

In this patch, we propagate the failure to the caller. This will allow
the caller to determine how to handle this - whether to fail, or swallow
the error and move on.

Bug: 434819179
Flag: EXEMPT refactor
Test: atest CtsInputTestCases
Test: atest CtsViewTestCases
Change-Id: I1b4206aeb1943134401845f9db224e9b75692498