Update color space extensions advertised to always be present when EGL_KHR_gl_colorspace is present.
Previously, the EGL loader used a combination of the board config for
wide color and HDR to decide what extensions it advertised. This has
been changed to always include the complete list of known color space
extensions when the base EGL_KHR_gl_colorspace extension is present.

This makes the loader match the behavior of SurfaceFlinger which will
correctly handle the color space selection if the wide color and HDR
board configs are not present.

Bug: b/435496982
Flag: EXEMPT small bugfix
Change-Id: Ie03ec18426509012a0302db7a72beb69dc47057a