rustBinderTest: avoid duplicate test names
When multiple tests start a helper process with the same name, they're
going to get mixed up with each other. This can lead to test flakiness
in case a test uses the service from the wrong test, and the sub-process
gets killed before the other test is done using it. Fix the service
names so that this doesn't happen.

If the call to synchronize_rcu() in the binder driver is changed to be
called unconditionally, then the resulting changes in timing mean that
this flake happens almost every time instead of only 1-2% of the time.
This makes it easier to test whether the issue has been fixed.

Flag: EXEMPT test only
Test: Run test with synchronize_rcu() called unconditionally in Binder
Change-Id: Ia74da4891261dc810cc9760acd3a62e05c23646a
Signed-off-by: Alice Ryhl <aliceryhl@google.com>