SF: fix DisplayTransactionCommitTest#processesVirtualDisplayAdded
This fixes this test by updating the expectations and properly setting
the origin of the sink surface. This only affects the unlaunched
VirtualDisplaySurface2 code.

This moves the construction of the render surface to onFirstRef so that
the surface can be accessed immediately after construction. Previously
this would be created only after the sink surface was set up.

Bug: 340933138
Flag: com.android.graphics.surfaceflinger.flags.wb_virtualdisplay2
Test: DisplayTransactionCommitTest

Change-Id: Ifcea02d5759065159a6ffce7d1d06a2adff3835b