SF: Update supported refresh rates on mode and frame rate override update
On ARR devices we return all the divisor rates, and when mode changes
we want to update the list with the new possible list of rates for
the api. This update will update the DisplayManager with the new list of
divisors.

Test: atest RefreshRateSelectorTest
Test: atest CtsSurfaceControlTests
Test: atest EventThreadTest
BUG: 431045537
Flag: com.android.graphics.surfaceflinger.flags.supported_refresh_rate_update
Change-Id: Id755b0553a26f53937b57b3c8952eec3a86bb1e4