Refactor BBinder to move mRecordingOn into Extras
This change moves the mRecordingOn flag from the BBinder object itself
into the heap-allocated Extras struct. Since this is a debug-only
feature, it is acceptable for it to cause an allocation.

This refactoring frees up a bit in the BBinder object for other flags
and has no functional impact.

Bug: 394661050
Test: atest binderRecordReplayTest
Flag: EXEMPT refactor
Change-Id: I1677d231f3a138dc6ebb77a6884e57463d02d8a2